code/requirements.txt
matplotlib==3.4.2
numpy==1.19.2
Pillow==8.4.0
scikit_image==0.17.2
skimage==0.0
torch==1.7.1
torchvision==0.8.2
